osbuild-worker: split out worker.go
No functional change.
